requirements.txt
asgiref==3.5.2
Django==4.0.5
django-versatileimagefield==2.2
djangorestframework==3.13.1
drf-flex-fields==0.9.8
Markdown==3.3.7
Pillow==9.1.1
python-magic==0.4.27
pytz==2022.1
sqlparse==0.4.2
